Move boss cmd tests to their own file
